Extract state management about protocol to separate file, mostly
Still need to move `rpc`, which is going to require channels
