+++ magic-wormhole.cabal
@@ -35,6 +35,7 @@ library
     , unordered-containers
     , websockets
   exposed-modules:
+      MagicWormhole.Internal.Dispatch
       MagicWormhole.Internal.Messages
       MagicWormhole.Internal.Rendezvous
       MagicWormhole.Internal.WebSockets
